[#snag-013-shortcut-for-open-issue.012-rename-to-snag]
  + (DONE rename api 'issue' to 'node' too) (step 2 of 2)

+++ lib/skylab/snag/test/cli/actions/add_spec.rb
@@ -30,7 +30,11 @@ module Skylab::Snag::TestSupport::CLI::Actions
         client_invoke 'add', '-n', '-v', 'foo bizzle'
       end
       o( / new line: / )
-      o( / rm / )
+      if output.lines.first.string =~ /mkdir .+snag-PROD/
+        output.lines.shift # egads sorry this is bad
+      else
+        o( / rm / )
+      end
       o( / mv /)
       o( / mv /)
       o( / done\./ )
